fix: remove duplicate function
