+++ store/src/scoring.rs
@@ -1,4 +1,7 @@
-use crate::dto::{BlacklistRecord, ScoringRunRecord, UnstakeHint, UnstakeHintRecord};
+use crate::dto::{
+    BlacklistRecord, ScoringRunRecord, UnstakeHint, UnstakeHintRecord, ValidatorScoreRecord,
+};
+use chrono::{DateTime, Utc};
 use rust_decimal::prelude::*;
 use std::collections::{HashMap, HashSet};
 use tokio_postgres::Client;
@@ -84,10 +87,7 @@ async fn voters_credits_performance_in_epoch(
     psql_client: &Client,
     epoch: u64,
 ) -> anyhow::Result<HashMap<String, f64>> {
-    log::info!(
-        "Loading list of poor voters: {}",
-        epoch
-    );
+    log::info!("Loading list of poor voters: {}", epoch);
     Ok(psql_client
         .query(
             "WITH stats AS (SELECT AVG(activated_stake * credits) / avg(activated_stake) as stake_weighted_avg_credits FROM validators WHERE epoch = $1)
@@ -121,7 +121,8 @@ pub async fn load_unstake_hints(
     } else {
         Default::default()
     };
-    let voters_credits_performance = voters_credits_performance_in_epoch(psql_client, epoch).await?;
+    let voters_credits_performance =
+        voters_credits_performance_in_epoch(psql_client, epoch).await?;
     let blacklist = load_blacklist(blacklist_path)?;
 
     for (vote_account, commission) in commissions_in_this_epoch {
@@ -171,6 +172,82 @@ pub async fn load_unstake_hints(
             },
         )
         .collect())
+}
+
+pub async fn load_all_scores(
+    psql_client: &Client,
+) -> anyhow::Result<HashMap<String, Vec<ValidatorScoreRecord>>> {
+    log::info!("Querying all scores...");
+    let rows = psql_client
+        .query(
+            "
+            SELECT vote_account,
+                score,
+                rank,
+                mnde_votes,
+                ui_hints,
+                component_scores,
+                component_ranks,
+                component_values,
+                eligible_stake_algo,
+                eligible_stake_mnde,
+                eligible_stake_msol,
+                target_stake_algo,
+                target_stake_mnde,
+                target_stake_msol,
+                scores.scoring_run_id,
+                scoring_runs.created_at as created_at
+            FROM scores
+            LEFT JOIN scoring_runs ON scoring_runs.scoring_run_id = scores.scoring_run_id
+            ORDER BY rank",
+            &[],
+        )
+        .await?;
+
+    let records: HashMap<_, Vec<_>> = {
+        log::info!("Aggregating scores records...");
+        let mut records: HashMap<_, Vec<_>> = Default::default();
+        for row in rows {
+            let vote_account: String = row.get("vote_account");
+            let scores = records
+                .entry(vote_account.clone())
+                .or_insert(Default::default());
+            scores.push(ValidatorScoreRecord {
+                vote_account: vote_account.clone(),
+                score: row.get("score"),
+                rank: row.get("rank"),
+                mnde_votes: row.get::<_, Decimal>("mnde_votes").try_into().unwrap(),
+                ui_hints: row.get("ui_hints"),
+                component_scores: row.get("component_scores"),
+                component_ranks: row.get("component_ranks"),
+                component_values: row.get("component_values"),
+                eligible_stake_algo: row.get("eligible_stake_algo"),
+                eligible_stake_mnde: row.get("eligible_stake_mnde"),
+                eligible_stake_msol: row.get("eligible_stake_msol"),
+                target_stake_algo: row
+                    .get::<_, Decimal>("target_stake_algo")
+                    .try_into()
+                    .unwrap(),
+                target_stake_mnde: row
+                    .get::<_, Decimal>("target_stake_mnde")
+                    .try_into()
+                    .unwrap(),
+                target_stake_msol: row
+                    .get::<_, Decimal>("target_stake_msol")
+                    .try_into()
+                    .unwrap(),
+                scoring_run_id: row.get("scoring_run_id"),
+                created_at: row
+                    .get::<_, DateTime<Utc>>("created_at")
+                    .try_into()
+                    .unwrap(),
+            })
+        }
+
+        records
+    };
+    log::info!("Records prepared...");
+    Ok(records)
 }
 
 pub async fn load_scoring_runs(psql_client: &Client) -> anyhow::Result<Vec<ScoringRunRecord>> {
